Added Find functionality
The Find Page now works and shows a table displaying the syllabus you searched for. You can print to PDF and delete a syllabus. HOWEVER, since Find Page is a function, when you delete a syllabi, it does not update automatically, so you have to refresh and search again to see the change. This would be fixed if it was a class.

Also cleaned up CSS for Home Page and Find Page.