Fixed Scrolling Issue
I made it to where the Table display on the Homepage does not overflow on the main page itself so there should be no vertical scroll bar on the main page. The Table overflows in its own container so it should always have the horizontal and vertical scroll bars present.